wip: use the new aux data provider

+++ crates/asm/common/src/aux/errors.rs
@@ -85,7 +85,7 @@ pub enum ManifestLeavesError {
     #[error("invalid MMR proof at index {index}, hash {hash:?}")]
     InvalidMmrProofAtIndex {
         /// The index in the batch where verification failed
-        index: usize,
+        index: u64,
         /// The manifest hash that failed verification
         hash: Hash32,
     },
